refactor: created a Notfound component for reusability
